Support ranged archive retrieval
Glacier's retrieval cost is based on the size of the archive it needs
to fetch, spread over the four-hour-or-so retrieval window.  For a huge
archive this cost can be prohibitive, but by splitting the archive and
fetching it in chunks over a (likely much) longer period of time that
cost can be minimized, or eliminated completely.

Note, this currently requires a patched version of boto containing a
commit that hasn't been pushed upstream; see:
https://github.com/joshwilliams/boto